# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/me": {
      "get": {
        "operationId": "MeController_getMe",
        "parameters": [
          {
            "name": "x-unlock-token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Profile of the authenticated user",
        "tags": [
          "me"
        ]
      }
    },
    "/profile": {
      "put": {
        "operationId": "MeController_updateProfile",
        "parameters": [
          {
            "name": "x-unlock-token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Fill in / update the questionnaire (all fields required)",
        "tags": [
          "me"
        ]
      }
    },
    "/profile/photo": {
      "post": {
        "operationId": "MeController_uploadOwnPhoto",
        "parameters": [
          {
            "name": "x-unlock-token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "photo": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "photo"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Upload own photo (JPEG/PNG/WebP, ≤5 MB)",
        "tags": [
          "me"
        ]
      }
    },
    "/card": {
      "get": {
        "operationId": "MeController_getCard",
        "parameters": [
          {
            "name": "x-unlock-token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Authenticated journalist's current (primary) card",
        "tags": [
          "me"
        ]
      }
    },
    "/me/password": {
      "put": {
        "operationId": "MeController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change your own password",
        "tags": [
          "me"
        ]
      }
    },
    "/cards": {
      "get": {
        "operationId": "MeController_getCards",
        "parameters": [
          {
            "name": "x-unlock-token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "All cards the journalist holds (primary first)",
        "tags": [
          "me"
        ]
      }
    },
    "/card/primary": {
      "put": {
        "operationId": "MeController_setPrimary",
        "parameters": [
          {
            "name": "x-unlock-token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPrimaryCardDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Choose which card is primary",
        "tags": [
          "me"
        ]
      }
    },
    "/card/qr": {
      "get": {
        "description": "The client refreshes this every ~30 s; old QR codes stop verifying.",
        "operationId": "MeController_getCardQr",
        "parameters": [
          {
            "name": "cardId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-unlock-token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Fresh dynamic QR payload (short-lived signed verify URL)",
        "tags": [
          "me"
        ]
      }
    },
...
    "/admin/editorials": {
      "get": {
        "operationId": "EditorialsController_findAll",
        "parameters": [
          {
            "name": "x-unlock-token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List issuing companies (editorial admins: their own only)",
        "tags": [
          "admin/editorials"
        ]
      },
      "post": {
        "operationId": "EditorialsController_create",
        "parameters": [
          {
            "name": "x-unlock-token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEditorialDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create an issuing company (system admin only)",
        "tags": [
          "admin/editorials"
        ]
      }
    },
    "/admin/editorials/{id}": {
      "put": {
        "operationId": "EditorialsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "x-unlock-token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEditorialDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update an issuing company (editorial admins: their own only)",
        "tags": [
          "admin/editorials"
        ]
      },
      "delete": {
        "operationId": "EditorialsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete an issuing company (system admin only)",
        "tags": [
          "admin/editorials"
        ]
      }
    },
    "/admin/editorials/{id}/logo": {
      "post": {
        "operationId": "EditorialsController_uploadLogo",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "x-unlock-token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "logo": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "logo"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Upload the company logo (PNG/WebP/JPEG, ≤2 MB)",
        "tags": [
          "admin/editorials"
        ]
      }
    },
...
  },
  "info": {
    "title": "PressPass API",
    "description": "REST API for issuing, administration and verification of electronic press credentials.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "example": "Петренко Іван Васильович",
            "description": "ПІП"
          },
          "fullNameEn": {
            "type": "string",
            "example": "Ivan Petrenko",
            "description": "ПІП латиницею (для картки)"
          },
          "birthDate": {
            "type": "string",
            "example": "1990-05-15",
            "description": "Дата народження"
          },
          "passportData": {
            "type": "string",
            "example": "КВ 123456, виданий Шевченківським РВ 01.01.2010"
          },
          "taxNumber": {
            "type": "string",
            "example": "1234567890",
            "description": "ІПН — 10 цифр"
          },
          "phone": {
            "type": "string",
            "example": "+380501234567"
          }
        },
        "required": [
          "fullName",
          "birthDate",
          "passportData",
          "taxNumber",
          "phone"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "description": "Поточний пароль"
          },
          "newPassword": {
            "type": "string",
            "minLength": 8,
            "description": "Новий пароль"
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ]
      },
      "SetPrimaryCardDto": {
        "type": "object",
        "properties": {
          "cardId": {
            "type": "number",
            "example": 1,
            "description": "Id of the card to make primary"
          }
        },
        "required": [
          "cardId"
        ]
      },
...
      "CreateEditorialDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "ТОВ «Приклад Медіа»",
            "description": "Повна юридична назва"
          },
          "displayNameUk": {
            "type": "string",
            "example": "Онлайн-медіа «Приклад»",
            "description": "Назва для відображення в посвідченні (укр.)"
          },
          "displayNameEn": {
            "type": "string",
            "example": "«Pryklad» Media",
            "description": "Назва для відображення в посвідченні (англ.)"
          },
          "mediaId": {
            "type": "string",
            "example": "R40-02551",
            "description": "Ідентифікатор медіа (маска ***-*****)"
          },
          "edrpou": {
            "type": "string",
            "example": "12345678",
            "description": "Код ЄДРПОУ (8–10 цифр)"
          },
          "website": {
            "type": "string",
            "example": "https://pryklad.media/registry"
          },
          "director": {
            "type": "string",
            "example": "Іваненко Іван Іванович"
          },
          "email": {
            "type": "string",
            "example": "office@pryklad.media"
          },
          "address": {
            "type": "string",
            "example": "м. Київ, вул. Хрещатик, 1"
          },
          "phone": {
            "type": "string",
            "example": "+380441234567"
          },
          "cardNumberPrefix": {
            "type": "string",
            "example": "KV",
            "description": "Префікс нумерації посвідчень (унікальний серед редакцій)"
          },
          "cardNumberTemplate": {
            "type": "string",
            "example": "{prefix}-{year}-{seq:6}",
            "description": "Шаблон номера з токенів {prefix} {year} {YY} {seq} {seq:N} {mediaId}"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateEditorialDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "ТОВ «Приклад Медіа»",
            "description": "Повна юридична назва"
          },
          "displayNameUk": {
            "type": "string",
            "example": "Онлайн-медіа «Приклад»",
            "description": "Назва для відображення в посвідченні (укр.)"
          },
          "displayNameEn": {
            "type": "string",
            "example": "«Pryklad» Media",
            "description": "Назва для відображення в посвідченні (англ.)"
          },
          "mediaId": {
            "type": "string",
            "example": "R40-02551",
            "description": "Ідентифікатор медіа (маска ***-*****)"
          },
          "edrpou": {
            "type": "string",
            "example": "12345678",
            "description": "Код ЄДРПОУ (8–10 цифр)"
          },
          "website": {
            "type": "string",
            "example": "https://pryklad.media/registry"
          },
          "director": {
            "type": "string",
            "example": "Іваненко Іван Іванович"
          },
          "email": {
            "type": "string",
            "example": "office@pryklad.media"
          },
          "address": {
            "type": "string",
            "example": "м. Київ, вул. Хрещатик, 1"
          },
          "phone": {
            "type": "string",
            "example": "+380441234567"
          },
          "cardNumberPrefix": {
            "type": "string",
            "example": "KV",
            "description": "Префікс нумерації посвідчень (унікальний серед редакцій)"
          },
          "cardNumberTemplate": {
            "type": "string",
            "example": "{prefix}-{year}-{seq:6}",
            "description": "Шаблон номера з токенів {prefix} {year} {YY} {seq} {seq:N} {mediaId}"
          }
        }
      },
...
    }
  }
}